+++ framework/Message/Type.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Framework\Message;
+
+enum Type: string
+{
+    case Error = 'error';
+    case Warning = 'warning';
+    case Info = 'info';
+    case Success = 'success';
+}