+++ framework/Test/TestCase.php
@@ -2,8 +2,6 @@
 
 namespace Framework\Test;
 
-use AssertionError;
-
 /** Every unit test should base on this class. It provides the assertion functions. */
 class TestCase
 {
@@ -16,8 +14,27 @@ class TestCase
 
     public function assertEquals(mixed $expected, mixed $actual): void
     {
-        if (gettype($expected) !== gettype($actual) || $expected !== $actual) {
-            throw new AssertionFailedException($expected, $actual);
+        if (gettype($expected) !== gettype($actual)) {
+            throw new AssertionFailedException($this->mixedToString($expected), $this->mixedToString($actual));
         }
+
+        if (gettype($expected) === 'array') {
+            if (json_encode($expected) === json_encode($actual)) {
+                return;
+            }
+            throw new AssertionFailedException($this->mixedToString($expected), $this->mixedToString($actual));
+        }
+
+        if ($expected !== $actual) {
+            throw new AssertionFailedException($this->mixedToString($expected), $this->mixedToString($actual));
+        }
+    }
+
+    private function mixedToString(mixed $value): string
+    {
+        return match (gettype($value)) {
+            'array' => json_encode($value),
+            default => $value,
+        };
     }
 }