+++ framework/tests/Unit/Facades/TestConvert.php
@@ -5,7 +5,7 @@ use Framework\Test\TestCase;
 
 class TestConvert extends TestCase
 {
-    public function testBoolToInt()
+    public function testBoolToInt(): void
     {
         $this->assertEquals(1, Convert::boolToInt(true));
         $this->assertEquals(0, Convert::boolToInt(false));