+++ tests/Unit/TestTest.php
@@ -4,12 +4,12 @@ use \Framework\Test\TestCase;
 
 class TestTest extends TestCase
 {
-    public function testCase1()
+    public function testCase1(): void
     {
         $this->assertTrue(true);
     }
 
-    public function testCase2()
+    public function testCase2(): void
     {
         $this->assertTrue(true);
     }