+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Types of changes: `Added`, `Changed`, `Deprecate`, `Removed`, `Fixed`, `Secruity
 - Added routes for login and logout
 - Added command to create a new user
 - Added table for roles and user-role-assignment
+- Added message class to pass messages from controller to view
 
 ### Changed
 - Improved the autoloader to work for the new folder structure of version 2